backend/requirements.txt
django==5.1.2
psycopg2
djangorestframework==3.15.1
pillow==10.3.0
sqlparse==0.5.0
django-cors-headers